requirements.txt
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.6
Flask-Gravatar==0.5.0
Flask-Login==0.6.2
flask-sqlalchemy-report==0.2.2
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
Flask-WTF==1.0.1
Flask==2.1.2
Werkzeug==2.1.2
SQLAlchemy==1.4.40
WTForms==3.0.1
python-dotenv==0.20.0

